feat: ドキュメント更新通知機能を統合し、ベクトル化スクリプトを削除
- deploy.yml: Docusaurusキャッシュクリアのステップ名を変更し、通知とベクトル化のジョブを統合
- intro.md: postfix設定に関する詳細を追加
- notify-doc-updates.mjs: 不要なファイルを削除し、notify-and-vectorize.mjsに統合
- vectorize_articles.mjs: 不要なファイルを削除
- notify-and-vectorize.mjs: ドキュメントの更新を通知し、ベクトル化を行う新しいスクリプトを追加
